Update Rust to recent nightly
This isn't quite all the way to the latest internal version, because
that one introduces a new complicated problem, and this version is
already a good update.

Needed Updates:
* Update x86_64
* Apply changes to RK and stage0 due to x86_64 update
* Switch to naked_asm in naked functions (required now)
* Fix some clippy errors
  * Better lifetime elision
  * Suppress static mut denies for now
  * New c"" Cstr syntax
  * Doc/attribute hygiene
  * const -> static
  * clarify shifts with parens
  * inefficient bytes() usage
  * unnecessary map_err
  * using unit result
  * use repeat_n instead of repeat.take()
  * use next_back instead of last on double-ended tier

+++ oak_restricted_kernel/src/mm/page_tables.rs
@@ -53,7 +53,7 @@ pub unsafe fn create_offset_map<S: PageSize, M: Mapper<S>>(
         // page tables.
         mapper
             .map_to_with_table_flags(
-                Page::<S>::from_start_address(offset + i * (S::SIZE as usize)).unwrap(),
+                Page::<S>::from_start_address(offset + (i as u64) * S::SIZE).unwrap(),
                 frame,
                 flags,
                 PageTableFlags::PRESENT | PageTableFlags::WRITABLE | PageTableFlags::ENCRYPTED,
